requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.1.0
attrs==23.1.0
certifi==2023.11.17
discord.py==2.3.2
frozenlist==1.4.0
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
pyaes==1.6.1
pyasn1==0.5.0
PyYAML==6.0.1
rsa==4.9
sniffio==1.3.0
Telethon==1.32.1
yarl==1.9.2
